+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,12 +1,14 @@
 ## Before submitting
 
 - [ ] Was this discussed/approved via a Github issue? (no need for typos and docs improvements)
-- [ ] Did you read the [contribution guide](https://github.com/catalyst-team/catalyst/blob/master/CONTRIBUTING.md), Pull Request section?
+- [ ] Did you read the [contribution guide](https://github.com/catalyst-team/catalyst/blob/master/CONTRIBUTING.md)?
 - [ ] Did you check the code style? `catalyst-make-codestyle && catalyst-check-codestyle` (`pip install -U catalyst-codestyle`).
 - [ ] Did you make sure to update the docs? We use Google format for all the methods and classes.
 - [ ] Did you check the docs with `make check-docs`?
 - [ ] Did you write any new necessary tests?
-- [ ] If you made a notable change (that affects users), did you update the [CHANGELOG](https://github.com/catalyst-team/catalyst/CHANGELOG.md)?
+- [ ] Did you add your new functionality to the docs?
+- [ ] Did you update the [CHANGELOG](https://github.com/catalyst-team/catalyst/CHANGELOG.md)?
+- [x] **You can use 'Login as guest' to see Teamcity build logs.**
 
 <!-- For CHANGELOG separate each item in unreleased section by blank line to reduce collisions -->
 
@@ -36,8 +38,5 @@
 
 Anyone in the community is free to review the PR once the tests have passed.
 If we didn't discuss your PR in Github issues there's a high chance it will not be merged.
-
-**You can use 'Login as guest' to see Teamcity build logs.**
-
 
 <!-- Thank you for your contribution! -->
