+++ catalyst/contrib/dl/callbacks/cutmix_callback.py
@@ -142,3 +142,6 @@ class CutmixCallback(CriterionCallback):
                 * state.input[self.fields[0]].shape[-2]
             )
         )
+
+
+__all__ = ["CutmixCallback"]
