+++ catalyst/contrib/dl/callbacks/optimizer.py
@@ -87,3 +87,6 @@ class SaveModelGradsCallback(Callback):
 
             state.batch_metrics.update(**grad_norm)
             self._accumulation_counter = 0
+
+
+__all__ = ["SaveModelGradsCallback"]
