+++ catalyst/contrib/dl/callbacks/telegram_logger.py
@@ -126,3 +126,6 @@ class TelegramLogger(Callback):
                 )
 
                 self._send_text(text)
+
+
+__all__ = ["TelegramLogger"]
