+++ catalyst/contrib/models/cv/segmentation/README.md
@@ -1,6 +1,6 @@
 ## Segmentation contrib modules
 
-There are three main origins of inspiration for this contrib:
+Adapted from:
 1. [PyTorch-Toolbelt](https://github.com/BloodAxe/pytorch-toolbelt) by [BloodAxe](https://github.com/BloodAxe).
 2. [segmentation_models.pytorch](https://github.com/qubvel/segmentation_models.pytorch) by [qubvel](https://github.com/qubvel)
 3. [mmdetection](https://github.com/open-mmlab/mmdetection) by [open-mmlab](https://github.com/open-mmlab)
