+++ docs/api/data.rst
@@ -3,6 +3,13 @@ Data
 
 Data subpackage has data preprocessers and dataloader abstractions.
 
+.. toctree::
+   :titlesonly:
+
+.. contents::
+   :local:
+
+
 .. automodule:: catalyst.data
     :members:
     :undoc-members:
@@ -19,6 +26,7 @@ For example:
 
 .. automodule:: catalyst.data.__main__
     :members:
+    :exclude-members:    build_parser, main
 
 
 Augmentor
