+++ docs/api/dl.rst
@@ -10,16 +10,26 @@ DL
 Experiment
 --------------------
 
+.. automodule:: catalyst.dl.experiment
+    :members:
+    :show-inheritance:
+
+Experiment
+~~~~~~~~~~~~~~~~~~~~~~
 .. automodule:: catalyst.dl.experiment.core
     :members:
     :undoc-members:
     :show-inheritance:
 
+ConfigExperiment
+~~~~~~~~~~~~~~~~~~~~~~
 .. automodule:: catalyst.dl.experiment.config
     :members:
     :undoc-members:
     :show-inheritance:
 
+SupervisedExperiment
+~~~~~~~~~~~~~~~~~~~~~~
 .. automodule:: catalyst.dl.experiment.supervised
     :members:
     :undoc-members:
@@ -29,11 +39,19 @@ Experiment
 Runner
 --------------------
 
+.. automodule:: catalyst.dl.runner
+    :members:
+    :show-inheritance:
+
+Runner
+~~~~~~~~~~~~~~~~~~~~~~
 .. automodule:: catalyst.dl.runner.core
     :members:
     :undoc-members:
     :show-inheritance:
 
+SupervisedRunner
+~~~~~~~~~~~~~~~~~~~~~~
 .. automodule:: catalyst.dl.runner.supervised
     :members:
     :undoc-members:
@@ -43,22 +61,33 @@ Runner
 Callbacks
 --------------------
 
+.. automodule:: catalyst.dl.callbacks
+    :members:
+    :show-inheritance:
 
+ConfusionMatrixCallback
+~~~~~~~~~~~~~~~~~~~~~~~
 .. automodule:: catalyst.dl.callbacks.confusion_matrix
     :members:
     :undoc-members:
     :show-inheritance:
 
+InferCallback
+~~~~~~~~~~~~~~~~~~~~~~~
 .. automodule:: catalyst.dl.callbacks.inference
     :members:
     :undoc-members:
     :show-inheritance:
 
+MixupCallback
+~~~~~~~~~~~~~~~~~~~~~~~
 .. automodule:: catalyst.dl.callbacks.mixup
     :members:
     :undoc-members:
     :show-inheritance:
 
+LRFinder
+~~~~~~~~~~~~~~~~~~~~~~~
 .. automodule:: catalyst.dl.callbacks.scheduler
     :members:
     :undoc-members:
@@ -66,33 +95,45 @@ Callbacks
 
 
 Metrics
-~~~~~~~~~~~~~~~~
+--------------------
 
+Accuracy
+~~~~~~~~~~~~~~~~~~~~~~
 .. automodule:: catalyst.dl.callbacks.metrics.accuracy
     :members:
     :undoc-members:
     :show-inheritance:
 
+AUC
+~~~~~~~~~~~~~~~~~~~~~~
 .. automodule:: catalyst.dl.callbacks.metrics.auc
     :members:
     :undoc-members:
     :show-inheritance:
 
+Dice
+~~~~~~~~~~~~~~~~~~~~~~
 .. automodule:: catalyst.dl.callbacks.metrics.dice
     :members:
     :undoc-members:
     :show-inheritance:
 
+F1 score
+~~~~~~~~~~~~~~~~~~~~~~
 .. automodule:: catalyst.dl.callbacks.metrics.f1_score
     :members:
     :undoc-members:
     :show-inheritance:
 
+IOU
+~~~~~~~~~~~~~~~~~~~~~~
 .. automodule:: catalyst.dl.callbacks.metrics.iou
     :members:
     :undoc-members:
     :show-inheritance:
 
+Global precision, recall and F1-score
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 .. automodule:: catalyst.dl.callbacks.metrics.ppv_tpr_f1
     :members:
     :undoc-members:
@@ -102,16 +143,22 @@ Metrics
 Utils
 --------------------
 
+Torch
+~~~~~~~~~~~~~~~~~~~~~~
 .. automodule:: catalyst.dl.utils.torch
     :members:
     :undoc-members:
     :show-inheritance:
 
+Trace
+~~~~~~~~~~~~~~~~~~~~~~
 .. automodule:: catalyst.dl.utils.trace
     :members:
     :undoc-members:
     :show-inheritance:
 
+Wizard
+~~~~~~~~~~~~~~~~~~~~~~
 .. automodule:: catalyst.dl.utils.wizard
     :members:
     :undoc-members:
