+++ teamcity/py_deploy.sh
@@ -1,9 +1,13 @@
 #!/bin/bash
 set -e -o xtrace
 
-pip install -r requirements/requirements.txt
-pip install -r requirements/requirements-cv.txt
-pip install -r requirements/requirements-nlp.txt
+pip install \
+    -r ./requirements/requirements.txt \
+    -r ./requirements/requirements-dev.txt \
+    -r ./requirements/requirements-ml.txt \
+    -r ./requirements/requirements-cv.txt \
+    -r ./requirements/requirements-nlp.txt \
+    -r ./requirements/requirements-contrib.txt
 pip install -r docs/requirements.txt
 
 # @TODO: fix server issue
